+++ Source/GUI/Custom Components/EffectsTable.mm
@@ -57,7 +57,7 @@
 
 - (void)tableView:(UITableView *)tableView didSelectRowAtIndexPath:(NSIndexPath *)indexPath
 {
-    NSLog(@"%d", indexPath.row); // you can see selected row number in your console;
+    NSLog(@"%ld", (long)indexPath.row); // you can see selected row number in your console;
 }
 
 - (NSIndexPath *)tableView:(UITableView *)tableView willSelectRowAtIndexPath:(NSIndexPath *)indexPath {
